Getting things started.
Would love to say that I just made this exist how it is seen now.  That
would, of course, be a lie.  Worked on trying to get this to be a lot
more presentable for way longer than I'd care to admit.  I think more
time was spent making the document than the actual game code.  (Which
makes sense, but still...)